Replace run.sh by a custom installable script
Use setuptools to create an executable "spef" which executes the
project. The project must be installed to make the script available.

This required to add several non-python files to the installation
(spef/data/* and spef/*.yaml) using MANIFEST.in. The "spef/data" dir is
also important since log files are created inside it. In future all of
these files (config, logs, ...) should be placed outside of the package.