Migrate tests to pytest
It is more convenient to use a testing framework instead of hand-writing
the entire test execution.

The current tests are based on running SPEF and executing scenarios
(series of keystrokes) in it. Since pytest is by default capturing stdin
and stdout which prevents to use custom stdin, capturing is disabled.

The scenarios themselves are not changed.