Setup continuous integration using GitHub Actions
Run tests (pytest) and linter (black) in the CI.

The main problem is that GitHub Actions runners don't have a terminal
which is necessary for ncurses. This can be "hacked" by using the
'script' command which creates a pseudo-terminal. Also, ncurses requires
the TERM env variable to be set.

For now, we only test for Python 3.10, can add more versions in future.